+++ datas/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class DataDemoConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "data_demo"
+    name = "datas"
